+++ student.rb
@@ -1,11 +1,12 @@
-require_relative 'app'
+require './person.rb'
 
 class Student < Person
-  include Person
-  def initialize (classroam)
+  def initialize(name = 'Unknown', age, parent_permission: true)
+    super(name, age, parent_permission)
     @classroam = classroam
   end
+
   def play_hooky
-    puts "¯\(ツ)/¯"
+    "¯\(ツ)/¯"
   end
 end
